Break down big test file to get read for testing multiple server impls
