+++ package.json
@@ -15,6 +15,7 @@
     "@pulumi/cloud-aws": "^0.18.0",
     "@pulumi/pulumi": "latest",
     "@types/graphql": "^14.2.0",
+    "@types/lodash": "^4.14.123",
     "@types/node-fetch": "^2.3.2",
     "@types/uuid": "^3.4.4",
     "alsatian": "^2.4.0",
@@ -28,6 +29,7 @@
     "fp-ts": "^1.17.0",
     "graphql": "^14.2.1",
     "grip": "^1.3.0",
+    "lodash": "^4.17.11",
     "node-fetch": "^2.3.0",
     "tap-bark": "^1.0.0",
     "ts-dedent": "^1.0.0",
