+++ src/WebSocketApolloClient.ts
@@ -1,0 +1,48 @@
+import { InMemoryCache } from "apollo-cache-inmemory";
+import { ApolloClient, SubscriptionOptions } from "apollo-client";
+import { split } from "apollo-link";
+import { createHttpLink } from "apollo-link-http";
+import { WebSocketLink } from "apollo-link-ws";
+import { gql } from "apollo-server";
+import { getMainDefinition } from "apollo-utilities";
+import fetch from "node-fetch";
+import * as WebSocket from "ws";
+import { IApolloServerUrlInfo } from "./FanoutGraphqlExpressServer";
+
+const WebSocketApolloClient = ({
+  url,
+  subscriptionsUrl,
+}: IApolloServerUrlInfo) => {
+  const httpLink = createHttpLink({
+    fetch,
+    uri: url,
+    useGETForQueries: true,
+  });
+  const wsLink = new WebSocketLink({
+    options: {
+      reconnect: true,
+      timeout: 999999999,
+    },
+    uri: subscriptionsUrl,
+    webSocketImpl: WebSocket,
+  });
+  const link = split(
+    // split based on operation type
+    ({ query }) => {
+      const definition = getMainDefinition(query);
+      return (
+        definition.kind === "OperationDefinition" &&
+        definition.operation === "subscription"
+      );
+    },
+    wsLink,
+    httpLink,
+  );
+  const apolloClient = new ApolloClient({
+    cache: new InMemoryCache(),
+    link,
+  });
+  return apolloClient;
+};
+
+export default WebSocketApolloClient;
