--- src/subscriptions-transport-ws-over-http/package.json
@@ -1,63 +1,0 @@
-{
-  "name": "subscriptions-transport-ws",
-  "version": "0.9.16",
-  "description": "A websocket transport for GraphQL subscriptions",
-  "main": "dist/index.js",
-  "browser": "dist/client.js",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/apollostack/subscriptions-transport-ws.git"
-  },
-  "dependencies": {
-    "backo2": "^1.0.2",
-    "eventemitter3": "^3.1.0",
-    "iterall": "^1.2.1",
-    "symbol-observable": "^1.0.4",
-    "ws": "^5.2.0"
-  },
-  "scripts": {
-    "clean": "rimraf browser dist coverage",
-    "compile": "tsc",
-    "pretest": "npm run compile",
-    "test": "npm run testonly --",
-    "posttest": "npm run lint",
-    "lint": "tslint --format stylish --project ./tsconfig.json",
-    "watch": "tsc -w",
-    "testonly": "mocha --exit --reporter spec --full-trace ./dist/test/**/*.js",
-    "coverage": "node ./node_modules/istanbul/lib/cli.js cover _mocha -- --exit --full-trace ./dist/test/tests.js",
-    "postcoverage": "remap-istanbul --input coverage/coverage.raw.json --type lcovonly --output coverage/lcov.info",
-    "browser-compile": "webpack --config \"./unpkg-webpack.config.js\"",
-    "prepublishOnly": "npm run clean && npm run compile && npm run browser-compile"
-  },
-  "peerDependencies": {
-    "graphql": "^0.10.0 || ^0.11.0 || ^0.12.0 || ^0.13.1 || ^14.0.2"
-  },
-  "devDependencies": {
-    "@types/chai": "^4.0.0",
-    "@types/graphql": "^14.0.0",
-    "@types/is-promise": "^2.1.0",
-    "@types/lodash": "^4.14.109",
-    "@types/mocha": "^5.2.5",
-    "@types/node": "^8.0.8",
-    "@types/sinon": "^5.0.1",
-    "@types/ws": "^5.1.2",
-    "chai": "^4.0.2",
-    "graphql": "^14.0.2",
-    "graphql-subscriptions": "^1.0.0",
-    "istanbul": "^1.0.0-alpha.2",
-    "lodash": "^4.17.1",
-    "mocha": "^5.2.0",
-    "mock-socket-with-protocol": "^7.1.0",
-    "remap-istanbul": "^0.11.1",
-    "rimraf": "^2.6.1",
-    "sinon": "^6.1.4",
-    "tslint": "^5.10.0",
-    "typescript": "^2.9.1",
-    "webpack": "^3.1.0"
-  },
-  "typings": "dist/index.d.ts",
-  "typescript": {
-    "definition": "dist/index.d.ts"
-  },
-  "license": "MIT"
-}
